Wrote test for adding same email second time conflict, added 400 bad request for invalid status codes, minor refactorings
