+++ employ-easy/README.md
@@ -35,7 +35,7 @@ Use following command in terminal:  `mvn org.springframework.boot:spring-boot-ma
       1. Have your Docker Desktop up and running 
       2. Open your terminal (for example command prompt in Windows)
       3. Type `docker run --name postgresqldb -e POSTGRES_PASSWORD=postgres -d -p 5432:5432 postgres`
-      Press enter, make sure that you got a running container with the name "postres-challenges" in your Docker Desktop
+      Press enter, make sure that you got a running container with the name "postgres" in your Docker Desktop
       4.Type `docker ps` you will see the container id of your Docker container, copy it
       5.Type `docker exec -it b3bee3c04a62 bin/bash`, but put your container id instead of b3bee3c04a62
       6. Type `psql -U postgres`
