+++ employ-easy/src/main/java/se/yrgo/employee/controllers/EmployeeController.java
@@ -95,7 +95,7 @@ public class EmployeeController {
                     content = @Content(mediaType = "application/json",
                             schema = @Schema(implementation = EmployeeDTO.class))),
             @ApiResponse(responseCode = "400",
-                    description = "Resource with unique email cannot have value null", content = @Content),
+                    description = "Illegal employee/system status code or email was null", content = @Content),
             @ApiResponse(responseCode = "401",
                     description = "Authorization required to register the employee", content = @Content),
             @ApiResponse(responseCode = "403",
@@ -136,6 +136,8 @@ public class EmployeeController {
                     description = "Successfully updated an employee",
                     content = @Content(mediaType = "application/json",
                             schema = @Schema(implementation = EmployeeDTO.class))),
+            @ApiResponse(responseCode = "400",
+                    description = "Illegal employee/system status code or email was null", content = @Content),
             @ApiResponse(responseCode = "401",
                     description = "Authorization required to update the employee", content = @Content),
             @ApiResponse(responseCode = "403",
