+++ employ-easy/src/main/java/se/yrgo/employee/exceptions/ResourceExceptionHandler.java
@@ -1,11 +1,9 @@
 package se.yrgo.employee.exceptions;
 
-import javax.servlet.http.HttpServletRequest;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
-import se.yrgo.employee.exceptions.ObjectNotFoundException;
 
 @ControllerAdvice
 public class ResourceExceptionHandler {
@@ -28,6 +26,13 @@ public class ResourceExceptionHandler {
     public ResponseEntity<StandardError> nullPointer(NullPointerException e) {
         HttpStatus status = HttpStatus.BAD_REQUEST;
         StandardError err = new StandardError(status.value(), "Null Pointer", e.getMessage());
+        return ResponseEntity.status(status).body(err);
+    }
+
+    @ExceptionHandler(IllegalArgumentException.class)
+    public ResponseEntity<StandardError> illegalArgument(IllegalArgumentException e) {
+        HttpStatus status = HttpStatus.BAD_REQUEST;
+        StandardError err = new StandardError(status.value(), "Illegal Argument", e.getMessage());
         return ResponseEntity.status(status).body(err);
     }
 }
